Factor out invalid json response exception data provider
