+++ tests/Functional/Client/CreateFrontendTokenTest.php
@@ -7,13 +7,13 @@ namespace SmartAssert\UsersClient\Tests\Functional\Client;
 use GuzzleHttp\Psr7\Response;
 use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Message\ResponseInterface;
-use SmartAssert\UsersClient\Exception\InvalidResponseContentException;
-use SmartAssert\UsersClient\Exception\InvalidResponseDataException;
+use SmartAssert\UsersClient\Tests\Functional\DataProvider\InvalidJsonResponseExceptionDataProviderTrait;
 use SmartAssert\UsersClient\Tests\Functional\DataProvider\NetworkErrorExceptionDataProviderTrait;
 
 class CreateFrontendTokenTest extends AbstractClientTest
 {
     use NetworkErrorExceptionDataProviderTrait;
+    use InvalidJsonResponseExceptionDataProviderTrait;
 
     /**
      * @dataProvider networkErrorExceptionDataProvider
@@ -44,23 +44,6 @@ class CreateFrontendTokenTest extends AbstractClientTest
         $actual = $this->client->createFrontendToken('email', 'password');
 
         self::assertEquals($expected, $actual);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function invalidJsonResponseExceptionDataProvider(): array
-    {
-        return [
-            'invalid response content type' => [
-                'httpFixture' => new Response(200, ['content-type' => 'text/plain']),
-                'expectedExceptionClass' => InvalidResponseContentException::class,
-            ],
-            'invalid response data' => [
-                'httpFixture' => new Response(200, ['content-type' => 'application/json'], '1'),
-                'expectedExceptionClass' => InvalidResponseDataException::class,
-            ],
-        ];
     }
 
     /**
